requirements.txt
pyspark==3.5.1
jupyter
kafka-python
pandas
numpy
matplotlib
seaborn
streamlit
pymongo
altair
wordcloud